requirements/base.txt
Django>=3.2,<4.0
djangorestframework>=3.12.0
django-environ>=0.4.5
django-redis>=5.0.0
psycopg2-binary>=2.9.1
Pillow>=8.3.1
celery>=5.1.2
django-cors-headers>=3.7.0
django-rosetta>=0.9.8
drf-yasg>=1.20.0
PyJWT>=2.0.0
requests-oauthlib>=1.3.0
gunicorn==21.2.0 